requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
blinker==1.9.0
click==8.1.8
dill==0.3.9
exceptiongroup==1.2.2
Flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httptools==0.6.4
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
loguru==0.7.3
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.1.0
nonebot-adapter-onebot==2.4.6
nonebot2==2.4.1
pillow==11.0.0
pip==24.3.1
priority==2.0.0
propcache==0.2.1
pydantic==2.10.4
pydantic_core==2.27.2
pygtrie==2.5.0
python-dotenv==1.0.1
PyYAML==6.0.2
Quart==0.20.0
sniffio==1.3.1
toml==0.10.2
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
